requirements.txt
python-dotenv
langchain
langchain-core
langchain-google-genai
grpcio
streamlit
langchain-community
pypdf